requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
django-jazzmin==2.6.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
sqlparse==0.5.3
uritemplate==4.1.1
